+++ packages/koot/ReactSPA/client/run-first.js
@@ -1,1 +1,1 @@
-import '../../React/client-run-first'
+require('../../React/client-run-first')